+++ distribution/src/bin/elasticsearch-keystore.bat
@@ -3,7 +3,7 @@
 setlocal enabledelayedexpansion
 setlocal enableextensions
 
-set ES_MAIN_CLASS=org.elasticsearch.common.settings.KeyStoreCli
+set ES_MAIN_CLASS=org.elasticsearch.cli.keystore.KeyStoreCli
 set ES_ADDITIONAL_CLASSPATH_DIRECTORIES=lib/tools/keystore-cli
 call "%~dp0elasticsearch-cli.bat" ^
   %%* ^
